+++ Kernel/Net/E1000NetworkAdapter.cpp
@@ -5,6 +5,7 @@
  */
 
 #include <AK/MACAddress.h>
+#include <Kernel/Bus/PCI/API.h>
 #include <Kernel/Bus/PCI/IDs.h>
 #include <Kernel/Debug.h>
 #include <Kernel/Net/E1000NetworkAdapter.h>
