+++ Kernel/Net/RTL8168NetworkAdapter.cpp
@@ -5,6 +5,7 @@
  */
 
 #include <AK/MACAddress.h>
+#include <Kernel/Bus/PCI/API.h>
 #include <Kernel/Bus/PCI/IDs.h>
 #include <Kernel/Debug.h>
 #include <Kernel/Net/RTL8168NetworkAdapter.h>
