+++ Kernel/Storage/StorageManagement.cpp
@@ -5,6 +5,7 @@
  */
 
 #include <AK/UUID.h>
+#include <Kernel/Bus/PCI/API.h>
 #include <Kernel/Bus/PCI/Access.h>
 #include <Kernel/CommandLine.h>
 #include <Kernel/Devices/BlockDevice.h>
