+++ Userland/Utilities/lspci.cpp
@@ -70,7 +70,7 @@ int main(int argc, char** argv)
     VERIFY(json.has_value());
     json.value().as_array().for_each([db, format](auto& value) {
         auto& dev = value.as_object();
-        auto seg = dev.get("seg").to_u32();
+        auto domain = dev.get("domain").to_u32();
         auto bus = dev.get("bus").to_u32();
         auto device = dev.get("device").to_u32();
         auto function = dev.get("function").to_u32();
@@ -97,7 +97,7 @@ int main(int argc, char** argv)
         if (class_name.is_empty())
             class_name = String::formatted("{:02x}{:02x}", class_id, subclass_id);
 
-        outln(format, seg, bus, device, function, class_name, vendor_name, device_name, revision_id);
+        outln(format, domain, bus, device, function, class_name, vendor_name, device_name, revision_id);
     });
 
     return 0;
